Fix Gen64 for pre 1.8

+++ gopre18.go
@@ -4,5 +4,5 @@ package dummy
 
 func (g *UintGen) Gen64() uint64 {
 	diff := g.max - g.min
-	return g.min + (uint64(g.r.Uint32()<<32|g.r.Uint32()) % diff)
+	return g.min + (uint64(g.r.Uint32())<<32 | uint64(g.r.Uint32())%diff)
 }
